requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.4.0
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
certifi==2024.6.2
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.10
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.5.14
emoji==2.12.1
fastapi==0.99.1
filelock==3.15.4
filetype==1.2.0
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.0
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.23.4
humanfriendly==10.0
idna==3.7
importlib_resources==6.4.0
joblib==1.4.2
langchain==0.0.288
langsmith==0.0.92
lxml==5.2.2
Markdown==3.4.4
marshmallow==3.21.3
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.10.1
numpy==1.26.4
onnxruntime==1.18.0
openai==0.28.0
orjson==3.10.5
overrides==7.7.0
packaging==24.1
posthog==3.5.0
protobuf==5.27.1
pulsar-client==3.5.0
pydantic==1.10.17
PyPika==0.48.9
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-magic==0.4.27
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
starlette==0.27.0
sympy==1.12.1
tabulate==0.9.0
tenacity==8.4.2
tiktoken==0.7.0
tokenizers==0.19.1
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.2
unstructured==0.10.14
urllib3==2.2.2
uvicorn==0.30.1
watchfiles==0.22.0
websockets==12.0
yarl==1.9.4
